feat: 채팅 보낼 때 안읽음 수 계산

+++ src/main/java/com/sns/project/controller/chat/dto/response/ChatMessageResponse.java
@@ -15,9 +15,9 @@ public class ChatMessageResponse {
     private String sentAt;
     private Long senderId;
     private String senderName;
-    private int unreadCount;
+    private Long unreadCount;
 
-    public ChatMessageResponse(ChatMessage chatMessage, int unreadCount) {
+    public ChatMessageResponse(ChatMessage chatMessage, Long unreadCount) {
         this.id = chatMessage.getId();
         this.message = chatMessage.getMessage();
         this.sentAt = chatMessage.getSentAt().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
